+++ tests/test_cli_pitokenjanitor.py
@@ -1,0 +1,4 @@
+import unittest
+from flask.testing import CliRunner
+from privacyidea.cli import pi_token_janitor
+
